in progress

+++ admin/src/service/secret_service.rs
@@ -1,9 +1,9 @@
-use crate::helper::get_k8s_client;
 use crate::model::query_dto::{GatewayQueryDto, SgTlsQueryInst, ToInstance};
 use crate::model::vo::Vo;
 use crate::service::base_service::VoBaseService;
 use crate::service::gateway_service::GatewayVoService;
 use k8s_openapi::api::core::v1::Secret;
+use kernel_common::client::k8s_client;
 use kernel_common::constants::k8s_constants::DEFAULT_NAMESPACE;
 use kernel_common::{
     helper::k8s_helper::{format_k8s_obj_unique, parse_k8s_obj_unique, parse_k8s_unique_or_default, WarpKubeResult},
@@ -14,13 +14,15 @@ use kube::Api;
 use tardis::basic::error::TardisError;
 use tardis::basic::result::TardisResult;
 
+use super::spacegate_manage_service::SpacegateManageService;
+
 pub struct TlsVoService;
 
 impl VoBaseService<SgTls> for TlsVoService {}
 
 impl TlsVoService {
-    pub(crate) async fn list(clinet_name: &str, query: SgTlsQueryInst) -> TardisResult<Vec<SgTls>> {
-        let map = Self::get_type_map(clinet_name).await?;
+    pub(crate) async fn list(client_name: &str, query: SgTlsQueryInst) -> TardisResult<Vec<SgTls>> {
+        let map = Self::get_type_map(client_name).await?;
         if query.names.is_none() {
             Ok(map.into_values().collect())
         } else {
@@ -28,51 +30,50 @@ impl TlsVoService {
         }
     }
 
-    pub(crate) async fn add(clinet_name: &str, mut add: SgTls) -> TardisResult<()> {
-        #[cfg(feature = "k8s")]
-        {
+    pub(crate) async fn add(client_name: &str, mut add: SgTls) -> TardisResult<()> {
+        let is_kube = SpacegateManageService::client_is_kube(client_name).await?;
+        if is_kube {
             let (namespace, raw_nmae) = parse_k8s_unique_or_default(&add.get_unique_name());
             add.name = format_k8s_obj_unique(Some(&namespace), &raw_nmae);
         }
         let add_model = add.clone();
-        #[cfg(feature = "k8s")]
-        {
+        if is_kube {
             let (namespace, _) = parse_k8s_unique_or_default(&add.get_unique_name());
-            let secret_api: Api<Secret> = Self::get_secret_api(&Some(namespace)).await?;
+            let secret_api: Api<Secret> = Self::get_secret_api(client_name, &Some(namespace)).await?;
             let s = add_model.to_kube_tls();
             secret_api.create(&PostParams::default(), &s).await.warp_result_by_method("Add Secret")?;
         }
-        Self::add_vo(clinet_name, add).await?;
+        Self::add_vo(client_name, add).await?;
         Ok(())
     }
 
-    pub(crate) async fn update(clinet_name: &str, update: SgTls) -> TardisResult<()> {
+    pub(crate) async fn update(client_name: &str, update: SgTls) -> TardisResult<()> {
         let unique_name = update.get_unique_name();
-        if let Some(_old_str) = Self::get_str_type_map(clinet_name).await?.remove(&unique_name) {
-            #[cfg(feature = "k8s")]
-            {
+        let is_kube = SpacegateManageService::client_is_kube(client_name).await?;
+        if let Some(_old_str) = Self::get_str_type_map(client_name).await?.remove(&unique_name) {
+            if is_kube {
                 let (namespace, name) = parse_k8s_obj_unique(&unique_name);
-                let secret_api: Api<Secret> = Self::get_secret_api(&Some(namespace)).await?;
+                let secret_api: Api<Secret> = Self::get_secret_api(client_name, &Some(namespace)).await?;
                 let s = update.clone().to_kube_tls();
                 secret_api.replace(&name, &PostParams::default(), &s).await.warp_result_by_method("Update Secret")?;
             }
-            Self::update_vo(clinet_name, update).await?;
+            Self::update_vo(client_name, update).await?;
             Ok(())
         } else {
             Err(TardisError::not_found(&format!("[admin.service] Update tls {} not found", unique_name), ""))
         }
     }
 
-    pub(crate) async fn delete(clinet_name: &str, id: &str) -> TardisResult<()> {
-        #[cfg(feature = "k8s")]
-        {
+    pub(crate) async fn delete(client_name: &str, id: &str) -> TardisResult<()> {
+        let is_kube = SpacegateManageService::client_is_kube(client_name).await?;
+        if is_kube {
             let (namespace, name) = parse_k8s_obj_unique(id);
-            let secret_api: Api<Secret> = Self::get_secret_api(&Some(namespace)).await?;
+            let secret_api: Api<Secret> = Self::get_secret_api(client_name, &Some(namespace)).await?;
             secret_api.delete(&name, &DeleteParams::default()).await.warp_result_by_method("Delete Secret")?;
         }
-        let gateways = GatewayVoService::list(clinet_name, GatewayQueryDto { ..Default::default() }.to_instance()?).await?;
+        let gateways = GatewayVoService::list(client_name, GatewayQueryDto { ..Default::default() }.to_instance()?).await?;
         if gateways.is_empty() {
-            Self::delete_vo(clinet_name, id).await?;
+            Self::delete_vo(client_name, id).await?;
             Ok(())
         } else {
             Err(TardisError::bad_request(
@@ -85,7 +86,10 @@ impl TlsVoService {
         }
     }
 
-    async fn get_secret_api(clinet_name: &str, namespace: &Option<String>) -> TardisResult<Api<Secret>> {
-        Ok(Api::namespaced(get_k8s_client().await?, namespace.as_ref().unwrap_or(&DEFAULT_NAMESPACE.to_string())))
+    async fn get_secret_api(client_name: &str, namespace: &Option<String>) -> TardisResult<Api<Secret>> {
+        Ok(Api::namespaced(
+            (*k8s_client::get(client_name)).clone(),
+            namespace.as_ref().unwrap_or(&DEFAULT_NAMESPACE.to_string()),
+        ))
     }
 }
